+++ packages/react/src/exampleTypes/Text.js
@@ -1,5 +1,5 @@
 import _ from 'lodash/fp'
-import F from 'futil-js'
+import F from 'futil'
 import React from 'react'
 import { observer } from 'mobx-react'
 import { withNode, withLoader, withTreeLens } from '../utils/hoc'