+++ packages/react/src/greyVest/Style.js
@@ -1,5 +1,7 @@
 import React from 'react'
 
+let tableCellPadding = 8
+
 /*
 Component styles for the Grey Vest library. Each of these should be associated
 with a specific omponent in the lib. The css/emotion refactor will move each of
@@ -89,7 +91,7 @@ export default () => (
         border-bottom: solid 2px rgba(237, 237, 237, 0.43);
       }
       .gv-table td, .gv-table th {
-        padding: 20px;
+        padding: ${tableCellPadding}px;
         text-align: left;
       }
       .gv-table th > span {
@@ -205,31 +207,14 @@ export default () => (
 
 
       /* Tags Input */
-      .tags-input-one-line {
-        max-height: 40px;
-        overflow-y: auto;
-      }
-      .tags-input-one-line::-webkit-scrollbar {
-        -webkit-appearance: none;
-      }
-      .tags-input-one-line::-webkit-scrollbar-thumb {
-        border-radius: 8px;
-        border: 2.5px solid #f1f1f1; /* should match background, can't be transparent */
-        background-color: #c2c2c2;
-      }
-      .tags-input-one-line::-webkit-scrollbar-track {
-        background-color: #f1f1f1;
-        border-radius: 5px;
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
-      }
-      .tags-input > * {
-        box-sizing: border-box;
-      }
       .tags-input {
         border: 2px solid #EBEBEB;
         border-radius: 4px;
         background: #fff;
+      }
+
+      .tags-input > * {
+        box-sizing: border-box;
       }
       .tags-input input {
         height: 30px;