+++ packages/react/src/greyVest/Tag.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import F from 'futil-js'
+import F from 'futil'
 import { observer } from 'mobx-react'
 import Flex from './Flex'
 
@@ -19,6 +19,7 @@ let Tag = ({
   <span
     className="tags-input-tag"
     style={{
+      display: 'inline-block',
       cursor: 'pointer',
       margin: 3,
       borderRadius: '3px',