+++ packages/react/src/queryBuilder/QueryBuilder.js
@@ -1,13 +1,9 @@
 import React from 'react'
-import F from 'futil-js'
-import _ from 'lodash/fp'
-import { observer } from 'mobx-react'
+import F from 'futil'
 import DDContext from './DragDrop/DDContext'
 import Group from './Group'
 import styles from '../styles'
-import { withNode } from '../utils/hoc'
-import { withTheme } from '../utils/theme'
-import { useLens } from '../utils/react'
+import { contexturifyWithoutLoader } from '../utils/hoc'
 
 let { background } = styles
 
@@ -18,7 +14,7 @@ let QueryBuilder = ({
   mapNodeToProps,
   theme: { Button },
 }) => {
-  let adding = useLens(false)
+  let adding = React.useState(false)
   return (
     <div style={{ background }}>
       {node && (
@@ -40,11 +36,6 @@ let QueryBuilder = ({
   )
 }
 
-export default DDContext(
-  _.flow(
-    observer,
-    withNode,
-    withTheme
-  )(QueryBuilder),
-  { allowEmptyNode: true }
-)
+export default DDContext(contexturifyWithoutLoader(QueryBuilder), {
+  allowEmptyNode: true,
+})